Update get-config.ts
`ConnectionString` correctly replaces each `+` with space, because this is what the standard URL syntax is.

MongoDB decided to deviate from that, by treating `+` as just `+`. It is a bit a weird syntax, because they should have used `:` instead, i.e. like `mongodb:srv`, which is the proper way of doing it for the protocol.

This PR suggests that instead of patching the result from `ConnectionString`, to correct your expectation within `validProtocol` value, then you can just use it as is.

Note that the change will also make it work when the protocol is like `mongodb%20srv`, because `%20` is also a white-space in the URL ;)